# openapi.yaml
openapi: 3.0.0
info:
  title: Recipes rest - API
  description: Rest - API that allows users to obtain existing recipes data and post
    new ones.
  version: 1.0.0
servers:
- url: https://virtserver.swaggerhub.com/BGU-University/3.2_API/1.0.0
  description: SwaggerHub API Auto Mocking
...
paths:
...
  /auth/register:
    post:
      tags:
      - Auth
      summary: register user
      description: user register
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistration'
        required: true
      responses:
        "201":
          description: registration has been successfull.
        "409":
          description: invalid input
  /auth/login:
    post:
      tags:
      - Auth
      summary: login user
      description: user login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
        required: true
      responses:
        "200":
          description: login has been successfull.
        "405":
          description: invild input
  /auth/logout:
    get:
      tags:
      - Auth
      summary: login user
      description: user login
      responses:
        "200":
          description: succesfully logout.
        "405":
          description: Invalid Request - user not signed-in.
...
components:
  schemas:
...
    UserRegistration:
      type: object
      properties:
        username:
          type: string
          example: galated
        first_name:
          type: string
          example: Gal
        last_name:
          type: string
          example: Atedgi
        country:
          type: string
          example: Morroco
        password:
          type: string
        passwordAgain:
          type: string
        email:
          type: string
          example: galated@post.bgu.ac.il
        image:
          type: string
          example: some url...
...
    UserLogin:
      type: object
      properties:
        username:
          type: string
          example: galated
        password:
          type: string
...